+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 - LP Filter margin bottom
 - Disable the experimental feature 'Apply colors to global skin' by default
 - Only show title tab on object's root (If the permissions tab is available)
+- Support Certificate plugin
 
 ## [1.1.0]
 - View: Tile or list